[REFACTOR] Fix all pointer arithmetic
In C we use contiguous allocation and pointer arithmetics to build arrays. This
is managed automatically in Fortran and can be automated with operators in C++
but in C it remains manual.

- We don't use double**, we use double* and access the matrix indices with
pointer arithmetic.
- Storing row start indexes in a double** array for a regular matrix is not
useful, as pointer arithmetic operations are faster than memory lookup.
- We avoid pointer arithmetic operations (mat+i, or $mat[i]) unless they are
necessary, for instance when accessing a row.
- In this examples pointer arithmetic is only necessary to compile the buggy
matrix broadcast program.